+++ test/home_view_model_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:tano/features/notes/home_view_model.dart';
-import 'package:tano/core/repositories/notes_fixtures.dart';
 import 'package:tano/core/repositories/notes_repository.dart';
 import 'package:tano/core/models/note.dart';
 import 'package:tano/core/models/action.dart';
@@ -75,11 +74,6 @@ class _InMemoryNotesRepository implements NotesRepository {
     notes.clear();
   }
 
-  @override
-  Future<void> seedFixtures() async {
-    notes.clear();
-    notes.addAll(buildNotesFixtures());
-  }
 }
 
 Note _note(