+++ test/lock_feature_test.dart
@@ -3,7 +3,6 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:tano/core/models/note.dart';
-import 'package:tano/core/repositories/notes_fixtures.dart';
 import 'package:tano/core/repositories/notes_repository.dart';
 import 'package:tano/core/services/auth_service.dart';
 import 'package:tano/features/editor/edit_note_page.dart';
@@ -113,11 +112,6 @@ class _InMemoryNotesRepository implements NotesRepository {
     notes.clear();
   }
 
-  @override
-  Future<void> seedFixtures() async {
-    notes.clear();
-    notes.addAll(buildNotesFixtures());
-  }
 }
 
 Note _note({